+++ src/anemoi/datasets/create/functions/filters/rename.py
@@ -36,6 +36,7 @@ class RenamedFieldMapping:
         return getattr(self.field, name)
 
     def __repr__(self) -> str:
+        return repr(self.field)
         return f"{self.field} -> {self.what} -> {self.renaming}"
